requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asttokens==2.0.5
async-timeout==4.0.1
attrs==21.4.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
blis==0.7.4
boto3==1.20.24
botocore==1.23.24
brotlipy==0.7.0
cachetools==4.2.2
catalogue==2.0.6
certifi==2022.6.15
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.4
cryptography==36.0.0
cycler==0.11.0
cymem==2.0.5
daal4py==2021.5.0
DataProperty==0.55.0
debugpy==1.5.1
decorator==5.1.1
en-core-web-lg==3.1.0
en-core-web-sm==3.1.0
entrypoints==0.4
es-core-news-sm==3.1.0
executing==0.8.3
filelock==3.6.0
fonttools==4.25.0
frozenlist==1.2.0
future==0.18.2
gensim==4.2.0
google-api-core==1.25.1
google-auth==1.33.0
google-cloud-core==1.7.1
google-cloud-storage==1.31.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpcio==1.42.0
huggingface-hub==0.2.1
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.9.1
ipython==8.1.1
jedi==0.18.1
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.4.2
langcodes==3.3.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mbstrdecoder==1.1.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==5.2.0
munkres==1.1.4
murmurhash==1.0.5
nest-asyncio==1.5.4
nltk==3.7
numpy==1.21.2
packaging==21.3
pandas==1.2.3
parso==0.8.3
pathvalidate==2.5.0
pathy==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.4
preshed==3.0.5
prompt-toolkit==3.0.27
protobuf==3.19.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
PYEVALB==0.1.3
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.4
PySocks==1.7.1
pytablewriter==0.64.2
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2021.3
PyYAML==6.0
pyzmq==19.0.2
regex==2021.11.2
requests==2.27.1
rsa==4.7.2
s3transfer==0.5.0
sacremoses==0.0.43
scikit-learn==1.0.2
scikit-learn-intelex==2021.20220215.212715
scipy==1.7.3
sentencepiece==0.1.96
setuptools==58.0.4
shellingham==1.3.1
sip==4.19.13
six==1.16.0
sklearn-crfsuite==0.3.6
smart-open==5.1.0
spacy==3.1.6
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.1
stack-data==0.2.0
tabledata==1.3.0
tabulate==0.8.9
tcolorpy==0.1.2
thinc==8.0.13
threadpoolctl==2.2.0
tokenizers==0.11.4
torch==1.11.0
torchtext==0.12.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.18.0
typepy==1.3.0
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.8
wasabi==0.8.2
wcwidth==0.2.5
wheel==0.37.1
yarl==1.6.3
zipp==3.8.0
